backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
requests==2.31.0
openai==1.3.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-dateutil==2.8.2
email-validator==2.0.0

